requirements.txt
appdirs==1.4.3
astroid==1.4.9
Babel==2.3.4
colorama==0.3.7
Django==2.0.4
django-admin-list-filter-dropdown==1.0.1
django-autocomplete-light==3.2.10
django-bootstrap-form==3.2.1
django-bootstrap3==8.1.0
django-crispy-forms==1.6.1
django-debug-panel==0.8.3
django-debug-toolbar==1.6
django-filter==1.0.1
django-formtools==2.1
django-phonenumber-field==1.1.0
django-propeller==1.4.1
django-pyodbc-azure==2.0.4.1
django-suit==0.2.25
django-vanilla-views==1.0.4
djangorestframework==3.5.3
httplib2==0.9.2
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.8
mccabe==0.5.3
olefile==0.44
packaging==16.8
phonenumberslite==8.3.1
Pillow==4.0.0
psycopg2==2.6.2
Pygments==2.2.0
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
pyodbc==4.0.23
pyparsing==2.2.0
pytz==2016.10
six==1.10.0
sqlparse==0.2.2
virtualenv==15.1.0
wrapt==1.10.8
